+++ pysui/sui/sui_clients/common.py
@@ -75,8 +75,10 @@ class _ClientMixin(Provider):
     """
 
     _RPC_MINIMAL_VERSION: int = 19
-    # _SIGNATURE_ERROR: str = 'Invalid user signature: InvalidSignature { error: "signature error" }.'
-    _SIGNATURE_ERROR: str = "signature error"
+    _SIGNATURE_ERROR: set = {
+        'Invalid user signature: InvalidSignature { error: "signature error" }.',
+        "signature error",
+    }
 
     def __init__(self, config: SuiConfig, request_type: SuiRequestType = SuiRequestType.WAITFORLOCALEXECUTION) -> None:
         """Client initializer."""